refactor: split PabcService into thin PabcClient + domain logic in proxy
- PabcClient is now a thin HTTP client (auth, serialization only)
- Zaaktype filtering logic moved to ZaaktypeEnrichedProxyResult
- PabcConfig simplified to static constants holder
- Tests updated to match new client contract
